+++ public/app/features/dashboard/components/PanelEditor/OptionsPaneCategoryDescriptor.tsx
@@ -16,7 +16,7 @@ export interface OptionsPaneCategoryDescriptorProps {
   itemsCount?: number;
   customRender?: () => React.ReactNode;
   sandboxId?: string;
-  alwaysExpanded?: boolean;
+  isOpenable?: boolean;
 }
 
 /**
