+++ public/locales/pseudo-LOCALE/grafana.json
@@ -1324,6 +1324,7 @@
       "options": {
         "description": "Đęşčřįpŧįőŉ",
         "open-edit": "Øpęŉ Päŉęľ Ēđįŧ",
+        "plugin-type-image": "Ĩmäģę őƒ pľūģįŉ ŧypę",
         "title": "Päŉęľ",
         "title-option": "Ŧįŧľę",
         "transparent-background": "Ŧřäŉşpäřęŉŧ þäčĸģřőūŉđ"
